Update README.md
Readme File
The code provided to analyse the Reuters data set is pretty straightforward. You only need to understand the directions preceding the command.
Several algorithms need external package, and they must be download before running the code. They are mentioned at the beginning of the code.
Last thing, remember to select the right path of the Reuters data set. My, obviously, is different.  Also recall that some algorithms will take several minutes, as the data set and the number of features is not very small.